+++ resources/js/components/group/CreateGroups.vue
@@ -19,10 +19,13 @@
     </form>
     <div v-if="groups.length">
       <button class="button is-link" @click="createGroups(1)">
-        {{ trans.get("groups.new_group") }}
+        <i class="fas fa-plus mr-2"></i> {{ trans.get("groups.new_group") }}
       </button>
       <button class="button is-primary" @click="editGroups = !editGroups">
-        {{ trans.get("groups.edit_groups") }}
+        <i class="fas fa-edit mr-2"></i> {{ trans.get("groups.edit_groups") }}
+      </button>
+      <button class="button is-dark" v-if="studentsDnD() != 0" @click="random">
+        <i class="fas fa-random mr-2"></i> {{ trans.get("groups.random") }}
       </button>
       <drag-drop
         v-if="!editGroups"
@@ -62,6 +65,7 @@ import CustomCard from "./CustomCard.vue";
 
 export default {
   props: ["groups", "code", "students"],
+
   mounted() {
     this.csrfToken = document.querySelector('meta[name="csrf-token"]').content;
     let itemsDelete = [];
@@ -101,7 +105,32 @@ export default {
     DragDrop,
     CustomCard,
   },
-  methods: {
+  methods: { 
+    studentsDnD() {
+      if(this.$refs.dnditems)
+        return this.$refs.dnditems.items.length;
+      return false;
+    },
+    findSmallerGroup() {
+      let groups = _.shuffle(this.dropGroups);
+      let min = groups[0];
+      groups.forEach((element) => {
+        if(element.children.length < min.children.length)
+          min = element;
+      })
+      return min;
+    },
+    random() {
+      let students = _.shuffle(this.$refs.dnditems.items);
+      for (let i = 0; i < students.length; i++) {
+        let group = this.findSmallerGroup();
+        group.children.push(students[i]);
+        var index = this.$refs.dnditems.items.findIndex(function (item) {
+          return item.id === students[i].id;
+        });
+        this.$refs.dnditems.items.splice(index, 1);
+      }
+    },
     createGroups(groupsNumber = null) {
       if (!groupsNumber) groupsNumber = this.groupsNumber;
       axios
@@ -123,6 +152,7 @@ export default {
             type: "success",
           });
         });
+        this.$forceUpdate();
     },
     cancel() {
       location.reload();