+++ resources/lang/ca/students.php
@@ -91,6 +91,7 @@ return [
     'log_pet' => 'Mascota',
     'log_skill' => 'Resultat d\'una habilitat teva o d\'algú.',
     'log_task' => 'Has fet una tasca',
+    'log_volume' => 'Excès de soroll',
     
     'student_report' => 'Informe d\'estudiant',
 ];