+++ resources/lang/en/students.php
@@ -89,6 +89,7 @@ return [
     'log_pet' => 'Pet',
     'log_skill' => 'Skill result, yours or from another one.',
     'log_task' => 'You\'ve done one task',
+    'log_volume' => 'Volume exceded',
 
     'student_report' => 'Student report',